v1.0 Fixes
- Address selection of INA226's was incorrect, fixed by
simply moving 0 ohm resistors

- Instead of putting resistors in series with rs232 level translation
chip, they were put in series with the ttl connector. The resistors
should provide the device connected on the ttl connector to out
drive the level translation chip. Fix is to replace the resistors with
zero ohm resistors, and place them instead between the pad
and pin on the level translation chip.